Returns reference from locators and introduce testable interfaces
